Implement a notion of "operation classes"
When multiple operations share a prefix of arguments, this prefix can be
factored into a "superclass" which results in an intermediate C++ class.

The idea is that this can be used in LLPC to define a GenericLocationOp
class which has (location, elemIdx, etc.) arguments and is a C++ super
class of InputImportGenericOp, OutputExportGenericOp, etc.