Add Visitor and VisitorBuilder
Allow visitors to be written that refer to dialect operations. Visitors
are set up once in a manner similar to LLVM's TypeSwitch, but with an
interface that allows ahead-of-time optimizations if desired (like
building a hash table) and amortizing the cost of the optimization over
multiple uses of the visitor.