Remove the need for "dialect guards"
This is a quality of life improvement for users of the library, as shown
in practice by the problem underlying
https://github.com/GPUOpen-Drivers/llpc/pull/2244.

We still need to be able to map from LLVMContext to DialectContext
efficiently, in the face of multiple threads. To avoid always taking a
lock, each thread caches the most recent lookup in thread-local
storage. This shifts some cost (and complexity!) onto cleanup at context
destruction, but that's certainly worth it.

Kudos go to @jasilvanus for suggesting this approach.