Redesign constraint language and add DialectType
Add flexible DialectType support on top of LLVM's recent addition of
TargetExtTypes.

This pushes our ability to express argument and result constraints to
the limit, and most of this change is concerned with essentially
rewriting the constraint language as documented in docs/Constraints.md.

The language is now built around predicates. For verification, each
predicate must be satisfied given its arguments. But for result type
(and other) deduction, it is also possible to "evaluate" predicat
applications (compute the first argument from the rest) or "capture"
from them (compute another argument from the first one).

The ergonomics of the language revolve around UFCS, perhaps better
called "uniform predicate application syntax": in most cases, the first
or "self" argument to a predicate is implied by the position in which it
appears. During parsing, constraints are brought into a normal form
where the self argument is turned explicitly into a variable. A nest
like

  (FixedVectorType (IntegerType $bits), $count):$arg

is turned into

  (FixedVectorType $arg, $scalar_type, $count)
  (IntegerType $scalar_type, $bits)

There are more improvements that would be sensible, such as:

- Re-establish the use of Type::getInt{n}Ty for integer types that are
  cached in LLVMContext
- Cache evaluated sub-expressions in local C++ variables
- A few more language extensions like (cond) and (scope)

But those are left for later.

v2:
- fix the printable(...) implementation to do what it should
- address review comments