tablegen: explicitly represent builder methods
This simplifies the builder emission by splitting the decision logic
from the printing to some extent.

Also leverage the FmtContext substitutions feature a bit more
explicitly. This does tend to lead to nicer looking code.

At the same time, remove builderArgumentFilter: this was always a fairly
awkward feature that was only used to paper over non-opaque pointers.
Now that upstream LLVM and also LLPC are completing the switch to opaque
pointers, there's no good reason to keep it.

Finally, this now detects when the result type can be deduced from an
argument type via a SameType constraint.