Add ContextExtension
A context extension is a piece of extended data that can be attached to
a context. (Conceptually to an LLVMContext, in practice to a
DialectContext.)

This is a piece of infrastructure which would simplify the
implementation of DialectContext quite a bit if it were upstream.

We will use it in an upcoming verifier change to cache a Visitor object.