Add DialectExtension and DialectExtensionPoint
This is a generic mechanism that allows dialects to extend some
generically defined functionality.

Each dialect may or may not register an extension to each extension
point. Extensions are global (i.e., _not_ per-LLVMContext) and are
expected to be constant data.

For example, we will implement a generic verifier pass that verifies
dialect-specific aspects of an IR module (or function, etc.). To do so,
we will define a dialect extension point where each dialect registers an
extension for verifying the operations of that dialect.