Fix some issues with OpClass
OpClass allows a number of related operations to use a shared prefix of
arguments. The idea is to use this for some shader I/O operations in
LLPC that all need to talk about locations and components.

The support had regressed from my first steps in this direction due to
the lack of tests and the recent major changes, so let's add some basic
tests.