Allow unnamed types
Allow intrinsics with unnamed struct types.
The problem with unnamed struct types, is that mangling based on the
name does not work well, because multiple types map to the same mangled
name and getting a function with a mangled name then returns a
constantexpr bitcast of the existing function with the wrong type.

I see multiple ways to solve this:

1. After getting a function, check if it has the correct type and if it
   has the wrong type, append a suffix. Repeat until a function with the
   correct type is found or created.
   This is the easiest approach and implemented here.
2. Use a HashMap to map each function type to a name and use the map for
   fast lookups and to find unique suffixes.
   This shines when the first approach would have a lot of name
   collisions, but it needs memory and is the most complex approach to
   implement.
3. Same as 1., use linear probing in case of a name collision, but to
   reduce the chance of a conflict, choose the initial name with a
   random 2-digit suffix.

To test it, also add a test with typed pointers, which triggers a
different path when creating a function (with typed pointers, a bitcast
is returned; with opaque pointers, a function with a wrong type is
returned).