Visitor: Optimize core opcode and intrinsic ID lookup
Establish hash maps of core opcodes and intrinsic IDs that are being
visited and use those for the lookups instead of iterating over all
cases that have been added to the visitor.

No intended interface change, although this does potentially affect the
order in which callbacks are called, if multiple callbacks are installed
that can match the same instruction.

Callbacks installed for an explicit dialect op or intrinsic will always
be called before one that has been installed for the generic CallInst
core instruction type.

If multiple handlers for the same instruction become an issue, we may
want to allow some policies to be set in the future, including return
values from the callbacks indicating whether an instruction should be
processed further. This is still theoretical at the time of writing.