Visitor: Add convenient way of using member functions
Using member functions of the payload type seems likely to be a very
common pattern, especially for converting code that uses the
llvm::InstVisitor.

This complicates the VisitorHandler setup a bit. At some point, one
has to wonder if it wouldn't be better to just use std::function. The
current system still has some performance advantages:

* It is able to handle generic projections while still handling the
  common case of projections to fields cheaply, with a single, perfectly
  predictable branch.

* VisitorHandler is 32 bytes; so is std::function, so wrapping
  std::function would be more expensive in terms of cache pressure.

And we specifically do not want the full power of std::function. Users
of this facility should be strongly encouraged to use the pattern where
a visitor is constructed once as a static variable, so that the cost of
building hash maps is amortized.