tablegen: optimize the happy path in verifying logical OR expressions
Instead of unconditionally collecting the error messages for all
branches of a LogicOr and then throwing them all away if one of them was
successful, we generate the code for checking the branches twice.

For the initial check, no error messages are printed.

Then, only if all branches failed, the checks are run again, printing
error messages this time.