Fix vararg getter
arg_begin() and arg_end() return iterators over Use objects that are
stored contiguously in memory, but the Value objects that the uses point
to are very much not contiguous.

We solve this by introducing an iterator adaptor. The example code is
also adjusted to test this in a better way.

v2:
- use User::value_op_iterator