Simplify operating on sets and maps of dialect ops.
This commit introduces a new type, `OpSet`, that simplifies membership tests of
an instruction on a set of dialect operations. With this PR, applying a
callback on a given set of operations also becomes simpler since a visitor
function `addSet` is introduced.

It also introduces a map-like data structure, `OpMap`, which can be used
to simplify handling of association of dialect operations to certain
values.

With this PR, a new check-llvm-dialects-units target is added which
tests the new GTest ADT tests as well.