Remove typed pointers from example
Upstream llvm now asserts when trying to enable typed pointers.

llvm-dialects needs to keep support for typed pointers for some time,
but it also needs to compile with upstream llvm.

We could add a CMake flag or make the test code dependent on the llvm
version, but in practice, llvm-dialects examples are only tested here.