tablegen: relax a restriction of logic or
The constraint system evaluator refuses to evaluate a logic or unless
all variables that appear in it have been resolved. This is a good
thing, because we can't reliably capture variables from inside a logic
or.

However, the check was too restrictive since it also refused to evaluate
a logic or that contain a local/internal variable that is introduced
purely as a connective between "apply" constraints that appear as nested
tablegen DAGs.

We relax this restriction by making the distinction between "global" and
"non-global" variables explicit.