Add support for typed pointers with the builtin PointerType.
In some scenarios, we want to use typed pointers even if LLVM is moving
to opaque pointers. Dialects currently only supports opaque pointers in
its builtin. With this change, we simply return an i8* when evaluating a
builtin pointer type return value or operand. This translates to an opaque
pointer as well.
The client that creates the operations is responsible for
handling the pointers correctly (e. g. not using opaque pointers in a
typed pointer scenario).