Add VisitorResult
Allow a visitor callback to decide whether additional matching callbacks
for the same instruction should be called or not.

This is useful when a visitor callback wants to erase an instruction and
there are multiple visitors that might match the same instruction (as
sometimes happens with generic instructions like load and store).