Add support for immutable strings.
We want a way to store strings in a global variable by passing in a
StringRef to an op, but we don't want to generate a setter for it, since
we currently don't inject the builder into the setter. So, add an
immutable string type based on the existing isImmutable option for
attributes.